api-services/data-processing/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
pandas==2.1.4
numpy==1.24.3
pyodbc==5.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
azure-search-documents==11.4.0
openai==1.3.7
semantic-kernel==0.3.1
